+++ projects/_07_workforce_characteristics/_01_starters_leavers_vacancies/tests/fargate/utils/test_merge_utils.py
@@ -3,11 +3,6 @@ import projects._07_workforce_characteristics._01_starters_leavers_vacancies.far
 PATCH_PATH = "projects._07_workforce_characteristics._01_starters_leavers_vacancies.fargate.utils.merge_utils"
 
 
-class TestCreateListOfColsForAscwds:
-    def test_create_list_of_cols_for_ascwds(self):
-        pass
-
-
 class TestConvertAscwdsJobRoleColumnsToRows:
     def test_convert_ascwds_job_role_columns_to_rows(self):
         pass