+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes to this project will be documented in this file.
 
 ### Changed
 
+- Pull clean workplace columns into merge job within SLV pipeline.
 
 ### Fixed